examples/agentkit/requirements.txt
scalekit-sdk-python
python-dotenv
anthropic
protobuf
langchain-openai
langgraph
langchain-mcp-adapters
google-adk
crewai
crewai-tools
